Fix DelegateReader: Close the delegate reader only if it has not been created yet
This avoids swallowing exceptions in cases when the creation of the delegate reader will consume the entire contents of the inner reader, and close it. In that case a client of the delegate reader will call "close" which would trigger the creation of the delegate reader again, but the inner reader is already closed at that point, reporting IOException "stream closed" and hiding the original exception.